Added knownIssue as key for testament spec.
This is part of a series of commit with intent to re-organize the test suite.

Currently, the 'disabled' key of the testament spec is used both to disable tests on a platform basis and disable buggy tests.
Introducing the 'knownIssue' key in the spec allows simplifying the disabled key, and referencing knownIssues whose behaviour could ( and should ) be revisited.

Further discussion can be found at https://github.com/nim-works/nimskull/issues/61.

This commit adds a 'knownIssue' key for every test marked with a 'disabled: true' whose disabling could be linked to a knownIssue, be it from git-log messages, or test text.