debugutils - compiler analysis tracing
- use `addInNimDebugUtils` or create a specialized one for a proc you
  want to trace
- use `nimCompilderDebugRegion` along with the `nimDebugUtils` define
  in order to get traces for regions of code
- learn how the compiler does things... or doesn't as the case maybe

system has a template `nimCompilerDebugRegion` which is available when
compiling with `nimDebugUtils` defined. use this template as a block
and all code under it will then emit debug trace output from semExpr
and a number of other key procs within the compiler.

spam is kept to the point by creating delta stacktraces and nested
(indented) rendering of messages is used to convey the recursive nature
of this process.

the tracing facility is templatized and can be extended to add output
for additional situations if the present hooks do not suffice.

as part of this work the debugutils module was heavily gutted as it was
overwraught. there was an extension made to the compiler config to track
a small seq for delta stacktrace and output indentation tracking.

backstory:
- started with existing debugutils module, but it was over engineered
- removed most of the code and instead have a template to inject tracing

Future improvements:

- messages themselves need some love
- more things need traces