nimpretty/tester: always use built nimpretty
There are little reasons to use a nimpretty in PATH. The test suite
would only work for the nimpretty built from the same source.

This also removes the "feature" that automatically builds nimpretty
(which also depends on the nim in PATH, which might not be able to build
this at all). In the future this might return via koch.