[CHECKPOINT] Finished `extccomp.nim`
In total this module contained just about every possible way to write down
"command failed" message, **and** it used `errGenerated` (do we have errors
that are /not/ generated btw) - exactly the same kind of error that is used
in the ... sem, semexpr, parser, nimconf, linter, jsgen and so on. So much
for the structured logging.

- NEXT ::
  - I'm not sure, but `writeMapping` proc seems really weird, and by
    default I could not even get mapping file genrated. Potentially this
    data might be really useful, maybe report this information as well?