koch.py: support csources in release archives (#117)
This pull enables `koch.py` to build a bootstrap compiler from bundled
csources in release archives.

This allow a build from the source archive to be network-less.

Included is a small fix to ensure that `koch.py` is packaged in release
archives.

This pull also changes the source archive layout a fair bit, with all of
`csources` being placed in `/build/csources` instead of `/`.