This is the second commit of an ongoing effort to remove cruft from the test suite, and remove the 'disabled' keyword from the testament spec.
In the first commit, we added the new knownIssue key to the testament spec,
while also adding knownIssues to many of the tests for which we could attribute an issue.

In this commit, we almost finish the task of removing the 'disabled' keyword
from the remaining tests.

Below are the changes and reasonings behind them:

Removed Due to old or removed features:

    tdatabind.nim
    texplicit_dummy_closure.nim
    tgenconstraints.nim
    tincludefile.nim
    tmodulealias.nim
    tno_gcmem_in_shared.nim
    tnotuniquename2.nim
    tscheduler.nim

Removed due to being Half Baked or Non-sensical:

    tio.nim
    tmissingnilcheck.nim
    treadafterwrite.nim
    trecmod.nim
    tunsignedcomp.nim

Removed due to not being worth the trouble:

    t5000.nim

Removed due to being redundant:

    tisop.nim

Keeping disabled, due to being unsure:

    t7894.nim - Works, but flaky, and makes CI servers grumpy.
    tdllvar.nim
    tpragmas_misc.nim

Added knownIssues - Mostly textual descriptions, due to not having a explicit issues:

    tdont_init_instantiation.nim
    titerable.nim
    toutparam.nim
    toverload_asgn2.nim
    tparsopt.nim

Re-enabled:

    tglobal.nim
    tmultim.nim
    tmultimjs.nim
    tobjpragma.nim
    toverloading_typedesc.nim

Fixed

    trectuple.nim: Fixed the line number.
    trectype.nim : Fixed to the spec.