Remove NimblePkgList
I think it should be unbundled because is not required for a compiler to work.
No one is actually using it in production.
It is undocumented and not actively used.
It is not supported by Koch, you have to compile manually.
No relevant commits in a long time.
Less code to maintain.