testament: remove stdlib isMainModule testing
Presently testament will parse all modules under `pure/lib` and
`lib/packages/docutils` then:
1. it would read the contents, looking for a string match of
   `when isMainModule`
    1. resulting in a compile action instead of run
2. then include these modules as "tests" within its run

There are only two modules that are making use of this:
* `lib/packages/docutils/highlite.nim`: which acted as a build check
* `lib/pure/xmlparser.nim`: had a basic test case

For the former the build check was integrated into `tools/koch/kochDocs`
and for the latter a quick test was created within `tests/stdlib`.
After which point the `lib/packages/docutils` was removed from
testament's stdlib category.

Additionally, content parsing for `when isMainModule` has been removed,
as this is no longer necessary. This should speed up execution for slower
disks. Also these are now used as simple compiles tests instead of
attempting to run them as before. This at least ensures that pure modules
compile on all platforms.

The background motivation for this is it simplifies the testament rework
happen currently in: https://github.com/nim-works/nimskull/pull/135