text writer cleanup
Previous implementation [fn:1] used `writeLnHook` that were misused in
multiple contexts, including additional diagnositcs collection for the
sigmatch. It employed convoluted and rigid pipeline of message processing:

- Message was generated in-place, without any colored formatting
- It was passed down to the `liMessage` (removed now) procedure, that
  dispatched based on the global configuration state (whether
  `ignoreMsgBecauseOfIdeTools` holds for example)
- Then message, together with necessary bits of formatting (such as `Hint`
  prefix with coloring) was passed down to the
  `styledMsgWriteln(args: varargs[typed])` template, whcih did the final
  dispatch into
- One of the two different /macros/ for writing text out -
  `callStyledWriteLineStderr` and `callIgnoringStyle`.
  - Dispatch could happen into stdout/stderr or message writer hook

All of this convoluted madness is now replaced with

- `Report` object is instantiated
- `structuredErrorHook` converts it to string internally
- And calls `writeHook` to write out final string.

If colors need to be disabled, it is up to implementation of the
`structuredErrorHook` to provide necessary handing [fn:2].

[fn:1] This commit finalizes the change, it was finally made possible after
reports were properly structured.

[fn:2] previous implementation only used colors for `Hints` prefix (none of
the error message reports did any extra coloring, because why, people read
best when it is a monotone text all over the screen), but if we want to do
it /properly/, the colors might appear anywhere in the message (for example
to emphathise structured type diff - `seq[seq[int != string]]` on the type
mismatch errors.