refactor unittest.expect
current implementation of expect uses following features:

  * `body` argument after ``varargs[typed]``
  * exceptions for nnkExceptionBranch branch wrapped in nnkBracket

The argument after varargs is underspecified and badly implemented. In
other words it has bugs and it is unclear if they can be fixer or the
feature needs to be deprecated/removed.

Wrapping exceptions in nnkBracket for nnkExceptionBranch is in conflict
with astspec. According to the spec, exceptions must be added
directly to nnkExceptionBranch. It is questionable how and why it even
compiles.