Implement shell command runner wrapper
Most of the commands executed in the compiler go through
`execProcesses*(cmds: openArray[string], ...)`, or
`execCmdEx*(command: string, ...)` and as a consequence, most of the code
handles strings by immediately joining them. Leading to the usage of string
template interpolation all over the place.

While by itself it is not too bad, it makes it harder to reason about the code,
because semi-random string joins, potentially escaped (or forgotten), and
perhaps used in a hastily put together string interpolation system.

Closes https://github.com/nim-works/nimskull/issues/213