Implement variant object branch changing for `gc:refc`
This was previously forbidden and rejected at run-time. In order to get
`refc` closer to `arc`'s behaviour, it is now allowed.

The new logic is implemented in cgen and thus does not apply to the JS
back-end. It works by injecting reset logic for the assigned-to
discriminator's record-case in front of the assignment.

Behaviour is now the same as with `arc`, which means that:
* a branch reset now also happens when changing the discriminator value
  to a new value which refers to the same of-branch as the old one
* only the 'from' (source) branch is reset. This can lead to unexpected
  behaviour or crashes when modifying padding bits/bytes