cpp: remove cpp addr/deref handling in sem
In semexprs there was a path that changed addr/deref handling for cpp.
This is a layering violation and led to non-idempotent sem. Removal
will result in many code gen issues for the cpp backend.

This commit includes a number of changes to tests in order to ensure
that CI still passes. Identifying tests modified due to cpp related
failures with knownIssues.

Additionally, a few small issues with testament were found and fixed:
- tests weren't always skipped when they weren't in requested targets
- megatest always ran, even if the c target wasn't specified