compiler internals/api: qualify compiler imports
Compiler imports are now qualified with a `compiler` prefix. This
restores the ability to use the compiler API, previously resulted in
errors due to failed imports.

Due to the nature of this change many files had to be modified.

This particular fix was motivated in order to have `disruptek/nimph`
working under nimskull. Compiler API consumers can now use the API by
including `--path:"$nim"` in their cfg.

Misc:
- `tests/compilerapi/tcompilerapi` should catch regressions now
- doc work around, disabled example in `compiler/utils/nimdebugutils`,
  required because runnable examples doesn't account for the `path`