sem: use nkError in lookups.initOverloadIter
Converts old localReport to return an error symbol with an nkError ast.
Callsites were updated to check for error symbols. It's a little messy
but that this is part of the incremental refactoring necessary to have
sem become a simple function `context, untyped_ast -> typed_ast | error`

qualifiedLookup was also problematic as a localReport had snuck in there
it never should have made it in. The report was removed an audit of all
callsites were done to ensure they handled reporting duties as necessary.