vm: remove execution/code-gen inter-dependencies
Summary:
Move the logic for incremental execution-environment setup and
just-in-time code-generation  from `vmgen` into a separate module. This
removes all dependencies on and modifications of run-time resources
from code-generation.

All information required for executing a function (the bytecode offset
and register count) is now stored as part of it's function table entry,
allowing code-gen and subsequent execution to happen in fully separate
steps (a requirement for the first-class VM back-end).

In addition, `vmgen` no longer modifies the `position` field of any
symbol.

Misc:
- `VmFunctionObject` is renamed to `FuncTableEntry`, as the latter
 better reflects the type's purpose and usage
- the names at the usage sites of `FuncTableEntry` are adjusted