vm: remove lazy loading of `const` values
Summary:

Values for `const`s are now loaded into VM memory on their first
appearance instead of on their first access. This has no effect on
guest-visible semantics.

The responsibility of loading the constant value is moved outside the
execution engine, simplifying the latter and making the loading logic
easily interchangeable. In addition, it also make implementation of
the VM back-end simpler.

Details:

- no dedicated layer responsible for updating the execution
 environment with the dependencies exist yet, so the loading of the
 `const` values is temporarily done in `vmgen`