vm: change how string constants are stored
Store string VM constants as `string` instead of `VmString`.

This previous approach required a layering violation and an unwanted
dependency, since the VM's allocator (a run-time resource) needed for
creating `VmString`s was used and modified by the code-generator
(i.e. at compile-time).

As a consequence of this change, string literals usage (in both `var`
and `let` contexts) always means an additional allocation + copy for
now. Implementing support for foreign memory in the VM will provide
the framework to fix this.