Added 'labels' key in testament spec; Love for the test suite
Continuing with the theme of enhancing our test suite,
this commits adds another key to the testament spec.

The `labels` key is a space-separated string of metadata
describing the test, with purpose of being used
during development.

Since a full run of the test suite takes a long time,
it makes sense to run only a subset at a time. The
current `cat` folders does a very poor job , being
coarse and arbitrary. This key aims to lay a
foundation to fill the gaps.

We also continue with the efforts setted by the
previous commits:
. Remove almost all illegible names in files.
. Add a meaningful 'description' for the changed tests.
. Reduce the usage of 'echo' in the tests.

This is part of a bigger effort, see
https://github.com/nim-works/nimskull/issues/41.

Co-authored-by: zerbina <100542850+zerbina@users.noreply.github.com>