compiler/front: add depfile option
Add a new command-line option (--depfile) for the nim compile command.
It accepts an argument (output path for a depfile, generated for a given
target).

This change improves integrations with language-independent build
systems. Particularly, helps to get better edit-compile cycle times.

Previously a build system needed to rebuild a binary if any of source
files had been changed, even if they weren't used by the binary. The
full list of files that a given source file depends on can only be
discovered by the compiler.

It doesn't need to be a standalone command because if the file has never
been compiled, it must be built anyway, generating dependencies as a
side effect.

See PR #376 for more information and examples.