initial implementation of the descriptive `[]`
Initial implementation to test validity/usability of the more
descriptive accessor procedures: instead of writing `[0..^3]`
`SliceAllIdents` can be used.

Right now rough estimates of the `\[[0-9]+\]` usage total to 5498
instances in the `compiler/` directory, with 394 in VM, 2359 in `sem/`
etc.

This PR provides core part/implementation that should be expanded as it
is introduced in the different parts of the compiler.