vmir: better support for compiler-generated magic calls (part 1)
Instead of requiring magics calls to be represented via normal
procedure calls, they can now be represented directly with `ntkCall`
nodes without the need for a `ProcId`.

The old approach had a few problems. Since magics procedures can be
generic, it's not possible to know all valid instantiations once reaching
the backend. To still be able to insert magic calls, a pseudo-proc with
no type information was created for each magic and cached in `PassEnv`

This meant that `ProcHeader` together with all logic interacting with it
had to support procedures with no type information. In addition, the
cache for generated magics (`PassEnv`) had to be available wherever
magics need to be inserted.