compiler-magic handling follow-up (part 1)
The ability to use magics without requiring a `ProcedureEnv` entry
allows for a further optimization (and fix).

Right now, magic procedures having valid type information are still
registered with the environment. Not all of them are really valid
however, i.e. they're missing bodies, currently requiring guards against
all magic procedures in various places.

Depending on the used target/code-generator, some magics act as
normal procedures (they have a body and are not lowered), so the
aforementioned guards are also wrong, leading to the required
procedures not being generated.

The fix: all calls to magic procedures that are not treated as normal
ones are translated, during `irgen`, to use the dedicated `ntkCall`
variant for magics - not registering the called magic with the
environment. All procedures in the environment can now be
unconditionally treated as "real" procedures.

As magic calls are very common, this also reduces the memory usage,
since the previously required extra `ntkProc` node is now not needed
anymore. When compiling the compiler, this reduces the memory usage
by 4 MB. Less IR nodes also speeds up processing.

This part removes all usages of `genProcSym` in `genMagic` and is a
prerequisite for the actual change.