cbackend2/cpasses: use batch-processing for passes
Instead of sequentialy running and applying independent passes, they
are now processed in a batch. That is, they operate on the same input
IR and merge their changes into a `Changes` object which is then
applied to the IR after all passes in the batch are done.

This has a few advantages over the previous approach:
- less nodes to process, since the passes in a batch don't see the
 additions from each other
- temporarily semantically invalid IR is not visible to others passes in
 the batch, allowing for more flexibility
- the type information for the IR used by the batch needs to be computed
 only once