working prototype implementation for constants (part 1)
One important thing that is still missing from both the overall back-end
and code-generator is the handling of constants. Their
symbols/declarations are stored as part of `SymbolEnv`, but not their
corresponding literal data - the code-generator simply leaves their
definitions empty.

This series of commits implements a first version of support for them.
For simplicity (and to highlight why a dedicated IR makes sense), the
constants' data is still represented via `PNode` ASTs.

This commit adds the basics for storing the constants' data in
`SymbolEnv` plus the API for querying/modifying it.