working prototype implementation for constants (part 2)
Add the passes for transforming `set`, `string`, and `seq` constants.
Constants might themselves contain literals of those types, so an extra
pass for lifting these literals into constants is required.

To support more complex initializer expressions for constants, some node
kinds are - in the context of initializer expressions - overloaded with
extra meaning. Namely, these are:
- `nkRefTy`: a reference to a constant. In the generated code, the
 constant is to be embedded. Stores the ID of the referenced constant as
 an `nkIntLit` sub-node
- `nkPtrTy`: similar to `nkRefTy`, with the difference that the address
 it to be embedded instead
- `nkTupleConstr`: if used as the initializer for a record with a base
 type, the first element is the base record's initializer

The passes are not yet integrated into the processing pipeline.