introduce a dedicated IR for literal data
The IR distinguishes between four kinds of data:
- numbers (ints, uints, floats)
- strings
- packed integer arrays
- complex data (such as tuples, objects, and arrays/sequences)

Each of them uses a different internal representation. Numbers and
strings both use a `BiTable`, packed arrays use two sequences (one for
the raw data and one for the view information), and complex data uses a
flat packed-tree structure. Both packed arrays and complex data don't
use deduplication yet.

Since the data enters the back-end in the form of `PNode`s, a translation
step is necessary.

Instead of the previous `PNode`-based approach, constants now store their
corresponding literal data in the new IR. Both the transformation
pipeline for constants and the code-generator are adjusted to accomodate
for this change.