use `LiteralId` + `LiteralData` for all literals
`PNode` is now no longer used to represent any form of literal data.
This results in a large reduction in memory usage (for compiling the
compiler, by around 37 MB) - no improvement in
time-spent-in-backend is achieved however.

Due to all string-literals being present in the global `LiteralData`
object (thus allowing lookup via ID), the previously required
string-literal interning via `GlobalGenCtx.strings` can be (and is)
replaced with just using the string's `LiteralId`. In effect, this
allows for passing string literals coming out of `transf` through to the
code-generator without copies.

One significant downside of this change is that a `LiteralData` object
has to be available wherever one wants to add or inspect a literal,
making the interface of various procedures even more awkward than it
already is.

Most of the literals inserted during the various transformations are
currently untyped, so the code-generator has to support these for now.
This was much less of a problem with `PNode`, since `TNodeKind` is of
high-resolution compared to `LiteralKind`. E.g. without access
to a `TypeId`, nil literals, ints, uints, and float literals could still
be detected.