vmir: store `Literal`s as part of `IrNode3`
Instead of storing them in a separate sequence in `IrStore3`, they're
now stored directly in `IrNode3`. The reason behind this separation was
to have `IrNode3` free of GC'ed objects - `Literal` previously being one
(because of the `PNode`).

With the move to `LiteralId`, `Literal` is now no longer a GC'ed object,
so the separation is not necessary anymore. Due to `IrNode3` being a
variant object, a `Literal` is now also essentially "for free" in terms
of storage.