vm: properly handle `quit` in const/standalone mode
## Summary
The previous behaviour of `opcQuit` was to abort the execution loop
without any clean-up (and without reporting the user-provided
exit-code). Since this looked like a normal exit to the caller of
`rawExecute`, the VM was left in an invalid state, causing an internal
compiler error when `quit` was called in the context of a `const`
evaluation.

When being run in `emConst` or `emStandalone` mode, `opcQuit` now
performs an abnormal exit by raising an error that also stores the
exit-code.