apply the `seq` transform pass before the GC pass (part 1)
The `seq` transform pass and GC pass are currently heavily intertwined,
both doing transformations that should be done by the respective other
one. This is problematic for multiple reasons:
* it introduces multiple sources of truths
* it reduces modularity
* it tightly couples the seqsv1 implementation with an
  `unsureAsgnRef`-based GC

So instead, the passes are now adjusted to be able to run in separate
batches - the `seq` transform pass has to run before the GC pass
however.

Part 1 only implements the adjustments to the logic of both
passes - they're still run in the same batch.