apply the `seq` transform pass before the GC pass (part 2)
`seq`s are still lowered into `ptr` types, but the GC transform pass
requires `ref` types for the correct transform to be applied. For now,
the type-remapping feature of `TypeContext` is used in order to let
the GC transform pass know about `seq` and `string` types.

`lowerSeqTypesV1` now doesn't commit it's changes directly to the
environment but instead collects the mappings into a `TypeMap` which
can then be passed to `initTypeContext`.