cpasses: lower `ntkGotoLink` and `ntkContinue`
The first idea was to replace each `ntkGotoLink` with the section it
jumps to at the `CAst` level. The inlining approach (if done at the
`CAst` level) has some complexities regarding label naming, so instead,
the simpler solution of generating run-time dispatch logic was chosen.
This also results in smaller binaries.