cgen2: work around C type incompatibility
NimSkull `array`s are translated to C-arrays wrapped in a `struct`. This
is different from what `cgen` does and causes issues with imported types
and functions. As a quick-fix, if an expression of `array` type is used
in an argument context, it is first unwrapped.

This allows calls to `posix.pipe` to compile, which is required for the
compiler to bootstrap on systems that use the function.