cgen2: use `NU8` as the C type for `char`
Using `NIM_CHAR` (a macro for `char`) caused problems if the C-compiler
used signed chars, since NimSkull treats characters as unsigned.

`cgen` uses `rdCharLoc` (which casts the value to `NU8` if the location
is of char type) in various place in order to work around this issue.

A similar solution could be used here too - just using `NU8`
for `char` is much simpler however. `cstring` is changed to be
`unsigned char*`.