ast: remove `nkBreakState`
## Summary
`nkBreakState` is a compiler magic implemented via a `TNodeKind`, which
was originally implemented as "abort the current loop if the given iterator
has finished".

As part of #d71f69ab50f079c03860f244f6c64b555ca403b6, it was
changed to only expand to the "has finished" test, which, except for
additional logic for `nkClosure` is equivalent to what the implementation
of `system.finished` does.

In order to remove the redundancy, the `nkBreakState` node kind is
replaced with using the `mFinished` magic.

## Details
- generalize the `when defined` compatibility logic in `macros.nim` and
  move it into a template
- document `ccgstmts.genBreakState`
- adjust usage-sites of `mFinished` that expect the magic to be
  a normal procedure