compiler: remove support for hot-code-reloading
## Summary
HCR was only tested by a single test and made the code-generator
logic more complex. The feature is removed in order to reduce the
amount of code in `cgen` and to make future changes in the
code-generator easier.

The plan is to re-implement support for hot-code-reloading at a point
in the future.

## Details
- remove the HCR compiler support
- remove the HCR runtime
- remove the enabled test. The disabled `nimhcr_integration.nim` test is
  kept
- the `.nonReloadable` and `.executeOnReload` pragmas are removed