Sort toplevel callable and object entries
Only  toplevel  reordering   is  performed  as  there   are  many  internal
inconsistencies in directories themselves:  for example, `metatype` is used
for testing generic  objects, which also matches `generics`  in some cases.
There are tests that need to be  moved between different parts of the suite
later on. New structure:

```
lang_callable/
├── closure
├── converter
├── generics
│   └── sandwich
├── macros
├── method
├── overload
├── proc
├── procvar
├── template
└── varres

lang_objects/
├── constructors
├── destructor
├── fields
├── friends
├── init
├── metatype
├── objects
└── objvariant
```